backend/requirements.txt
# Framework Web
fastapi
uvicorn[standard]
python-multipart

# Variáveis de Ambiente
python-dotenv

# Processamento de Excel
openpyxl       # Arquivos .xlsx (Excel 2007+)
pandas         # Processamento de dados

# Database
mysql-connector-python

# OpenAI API
openai

# Autenticação e Segurança
bcrypt         # Hash de senhas
PyJWT          # JSON Web Tokens

# Testing
pytest
httpx  # Necessário para testes async do FastAPI
pytest-asyncio
httpx          # Para testar clientes HTTP async
